Move methods to model
